Moving files to correct structure
