+++ jestor/Table.py
@@ -1,6 +1,6 @@
 from typing import List
-from sdk.Client import Client
-from sdk.Filter.Filter import Filter
+from jestor.Client import Client
+from jestor.filter.Filter import Filter
 
 class Table:
     def __init__(self, token, org, table_name):
