+++ jestor/filter/Filter.py
