+++ jestor/filter/Operators.py
